requirements.txt
Flask
python-dotenv
pandas
Pillow
torch
transformers
torchvision
scikit-learn
sentence_transformers
telebot